Cleanup and added TODOs

+++ Network.cpp
@@ -1,9 +1,11 @@
 #include "Network.h"
 
+//TODO implement
 int Network::evaluate() {
     return 0;
 }
 
+//TODO implement
 void Network::sgd(int epochsCount, int batchSize, double learningRate) {
 
 }
@@ -15,30 +17,37 @@ std::vector<double> Network::feedForward(std::vector<double> input) {
     return input;
 }
 
+//TODO implement
 std::vector<double> Network::sigmoidPrime(const std::vector<double> &z) {
     return std::vector<double>();
 }
 
+//TODO implement
 double Network::sigmoidPrime(double z) {
     return 0;
 }
 
+//TODO implement
 double Network::sigmoid(double z) {
     return 0;
 }
 
+//TODO implement
 std::vector<double> Network::sigmoid(const std::vector<double> &z) {
     return std::vector<double>();
 }
 
+//TODO implement
 std::vector<double> Network::costDerivative(const std::vector<double> &output, int label) {
     return std::vector<double>();
 }
 
+//TODO implement
 void Network::applyMiniBatch(const std::vector<std::pair<std::vector<double>, int>> &miniBatch, double learningRate) {
 
 }
 
+//TODO implement
 std::pair<std::vector<std::vector<double>>, std::vector<double>>
 Network::backPropagate(const std::vector<double> &test, int label) {
     return std::pair<std::vector<std::vector<double>>, std::vector<double>>();
@@ -63,15 +72,15 @@ Network::Network(std::vector<std::pair<std::vector<double>, int>> trainingData,
             layerWeights[i][j].resize(layerSizes[i + 1]);
         }
     }
-    for (auto & layerBias : layerBiases) {
-        for (double & bias : layerBias) {
+    for (auto &layerBias : layerBiases) {
+        for (double &bias : layerBias) {
             bias = arthas(rGen);
         }
     }
 
-    for (auto & layerWeight: layerWeights) {
-        for (auto & weights: layerWeight) {
-            for (double & weight : weights) {
+    for (auto &layerWeight: layerWeights) {
+        for (auto &weights: layerWeight) {
+            for (double &weight : weights) {
                 weight = arthas(rGen);
             }
         }
@@ -91,7 +100,7 @@ std::vector<double> Network::dot(const std::vector<std::vector<double>> &m, cons
 
 std::vector<double> Network::sum(const std::vector<double> &l, const std::vector<double> &r) {
     std::vector<double> sum(l.size());
-    for (int i = 0; i < l.size(); ++i) {
+    for (int i = 0; i < (int) l.size(); ++i) {
         sum[i] = l[i] + r[i];
     }
     return sum;
